struktur-edringer og begynt å utvikle eksportering av statistikk
